feat(web): add matches list and detail page
